Remove auto insert mode for commit files
This was an interesting idea, but ultimately I was hitting "i" and
deleting it right away. Maybe if I was starting Vim fresh, but alas this
old dog won't learn this trick. Back to classic commit mode.

P.S. Even this commit had an extra "i" to delete at the beginning