Disable mouse in Command mode
In MinTTY, I need to be able to get a right click menu to access some
options. Disabling mouse events in Command mode does the trick without
being super annoying.

For some reason, entering NVIM in MinTTY through Linux SSH is causing
the terminal to resize. The only way to get the size back to the
containing window is to Right-Click -> Full Screen. UGH!